Emergent pairs optimization; Move ripser functions to simplexArrayList
