+++ Complex/simplexBase.cpp
@@ -86,12 +86,10 @@ std::vector<std::vector<std::pair<std::set<unsigned>,double>>> simplexBase::getA
 }
 
 std::vector<simplexBase::treeNode*> simplexBase::getAllCofacets(const std::set<unsigned>& simplex){
-	ut.writeLog(simplexType,"No get cofacets function defined");
-	std::vector<simplexBase::treeNode*> ret;
-	return ret;
+	return getAllCofacets(simplex, 0, std::unordered_map<treeNode*, unsigned>(), false);
 }
 
-std::vector<simplexBase::treeNode*> simplexBase::getAllCofacets(const std::set<unsigned>& simplex, bool checkEmergent, double simplexWeight){
+std::vector<simplexBase::treeNode*> simplexBase::getAllCofacets(const std::set<unsigned>& simplex, double simplexWeight, const std::unordered_map<treeNode*, unsigned>& pivotPairs, bool checkEmergent){
 	ut.writeLog(simplexType,"No get cofacets function defined");
 	std::vector<simplexBase::treeNode*> ret;
 	return ret;
