+++ Pipes/fastPersistence.cpp
@@ -16,16 +16,13 @@
 #include <unordered_map>
 #include <queue>
 #include "fastPersistence.hpp"
-#include "simplexBase.hpp"
 
 unionFind::unionFind(int n) : rank(n, 0), parent(n, 0) {
 	for(int i=0; i<n; i++) parent[i]=i;
 }
 
 int unionFind::find(int i){
-	
-	//
-	if(i == parent[i]) return i;
+	if(i == parent[i]) return i; //Found name of the component
 	parent[i] = find(parent[i]); //Path Compression
 	return parent[i];
 }
@@ -45,76 +42,11 @@ bool unionFind::join(int x, int y){ //Union by rank
 	return true;
 }
 
-binomialTable::binomialTable(unsigned n, unsigned k) : v(n+1, std::vector<long long>(k+1, 0)){ //Fast computation of binomials with precomputed table
-	v[0][0] = 1;
-	
-	
-	for(int i=1; i<=n; i++){
-		v[i][0] = 1;
-		for(int j=1; j<=k; j++){
-			v[i][j] = v[i-1][j-1] + v[i-1][j]; //Pascal's Rule
-			if(v[i][j] < 0) throw std::overflow_error("Binomial overflow");
-		}
-	}
-}
-
-long long binomialTable::binom(unsigned n, unsigned k){ //Return binomial coefficient
-	if(k>n) return 0;
-	return v[n][k];
-}
-
 // basePipe constructor
 fastPersistence::fastPersistence(){
 	pipeType = "FastPersistence";
 	return;
 }
-
-//Hash the set by converting to a remapped index
-long long fastPersistence::ripsIndex(std::set<unsigned>& simplex, binomialTable& bin){ 
-	long long simplexIndex = 0;
-	unsigned i = 0;
-	auto it = simplex.begin();
-	while(it != simplex.end()){
-		simplexIndex += bin.binom(*it - shift, ++i);
-		if(simplexIndex < 0) throw std::overflow_error("Binomial overflow");
-		++it;
-	}
-	return simplexIndex;
-}
-
-unsigned fastPersistence::maxVertex(long long ripsIndex, unsigned high, unsigned low, unsigned k, binomialTable &bin){
-	while(high > low){ //Binary search for the max vertex for this simplex
-		unsigned mid = (high + low)/2;
-		if(bin.binom(mid, k) <= ripsIndex) low = mid + 1;
-		else high = mid;
-	}
-	return high - 1;
-}
-
-std::vector<unsigned> fastPersistence::getVertices(long long ripsIndex, int dim, unsigned n, binomialTable &bin){
-	std::vector<unsigned> v;
-	for(unsigned k = dim+1; k>0; k--){ //Get all vertices by repeated binary search for max vertex
-		n = maxVertex(ripsIndex, n, k-1, k, bin);
-		v.push_back(n);
-		ripsIndex -= bin.binom(n, k);
-	}
-	return v;
-}
-
-struct cmpBySecond{ //sort nodes by weight, then by lexicographic order
-	bool operator()(simplexBase::treeNode* a, simplexBase::treeNode* b) const{
-		if(a->weight == b->weight){ //If the simplices have the same weight, sort by reverse lexicographic order for fastPersistence
-			auto itA = a->simplex.rbegin(), itB = b->simplex.rbegin();
-			while(itA != a->simplex.rend()){
-				if(*itA != *itB) return *itA < *itB;
-				++itA; ++itB;
-			}
-			return false;
-		} else{
-			return a->weight > b->weight;
-		}
-	}
-};
 
 // runPipe -> Run the configured functions of this pipeline segment
 //
@@ -141,8 +73,8 @@ pipePacket fastPersistence::runPipe(pipePacket inData){
 	//Get all dim 0 persistence intervals
 	//Kruskal's minimum spanning tree algorithm
 	//		Track the current connected components (uf)
-	//		Check edges; if contained in current set ignore
-	//			if joins multiple sets, join all sets
+	//		Check edges; if contained in a component, ignore
+	//			if joins components, add them
 	//		Until all edges evaluated or MST found (size - 1)	
 	
 	std::priority_queue<unsigned> pivots; //Store identified pivots
@@ -170,7 +102,7 @@ pipePacket fastPersistence::runPipe(pipePacket inData){
 			inData.bettiTable.push_back(des);
 		}
 
-		//Check if we've filled our MST and can break...
+		//Check if we've filled our MST and can break
 		if(mstSize >= edges[0].size()-1) break;
 
 		edgeIndex++;
@@ -189,89 +121,34 @@ pipePacket fastPersistence::runPipe(pipePacket inData){
 	//	
 		//Build next dimension of ordered simplices, ignoring previous dimension pivots
 		
-		//Represent the ordered simplices as indexed sets; similar to the approach in indSimplexTree
+		//Represent the ordered simplices as indexed sets
 		
 		//Identify apparent pairs - i.e. d is the youngest face of d+1, and d+1 is the oldest coface of d
-		//		This indicates a feature represents a persistence interval
+		//		This indicates a feature represents a trivial persistence interval
 		
 		//Track V (reduction matrix) for each column j that has been reduced to identify the constituent 
 		//		boundary simplices
 	
-	// auto edge = edges[1][10];
-	// for(auto i : inData.complex->getAllCofacets(edge.first)){
-	// 	for(auto j : i->simplex) std::cout<<j<<' ';
-	// 	std::cout<<'\n';
-	// }
-
-	// if(true){
-	// 	for(unsigned d = 1; d < dim; d++){
-	// 		for(unsigned columnIndex = edges[d].size(); columnIndex-- != 0; ){
-	// 			std::pair<std::set<unsigned>, double>& simplex = edges[d][columnIndex];
-	// 			((simplexTree*)inData.complex)->getAllCofacets(simplex.first);
-	// 		}
-	// 	}
-	// } else{
-	
 	for(unsigned d = 1; d < dim && d < edges.size()-1; d++){
-		std::priority_queue<unsigned> nextPivots; 					//Pivots for the next dimension
-
-		std::unordered_map<unsigned, std::vector<unsigned>> v;			 //Store only the reduction matrix V and compute R implicity
-		std::unordered_map<simplexBase::treeNode*, unsigned> pivotPairs; //For each pivot, which column has that pivot
+		//TODO - Fix nextPivots for higher dimensional homology
+		std::priority_queue<unsigned> nextPivots;	 						//Pivots for the next dimension
+		std::unordered_map<unsigned, std::vector<unsigned>> v;				//Store only the reduction matrix V and compute R implicity
+		std::unordered_map<simplexBase::treeNode*, unsigned> pivotPairs;	//For each pivot, which column has that pivot
 
 		//Iterate over columns to reduce in reverse order
 		for(unsigned columnIndex = edges[d].size(); columnIndex-- != 0; ){ 
 			
-			std::pair<std::set<unsigned>, double>& simplex = edges[d][columnIndex];		//Pointer to the current simplex
+			std::pair<std::set<unsigned>, double>& simplex = edges[d][columnIndex];		//The current simplex
 			double simplexWeight = simplex.second;										//Current simplex weight
-			bool foundEmergentCandidate = false; 	//Found the lexicographically maximum cofacet with the same diameter
 
 			//Not a pivot -> need to reduce
 			if(pivots.top() != columnIndex){ 
-				std::vector<simplexBase::treeNode*> cofaceList = inData.complex->getAllCofacets(simplex.first);
-				std::vector<unsigned> columnV;
-				columnV.push_back(columnIndex);
+				//Get all cofacets using emergent pair optimization
+				std::vector<simplexBase::treeNode*> cofaceList = inData.complex->getAllCofacets(simplex.first, simplexWeight, pivotPairs);
+				std::vector<unsigned> columnV;	//Reduction column of matrix V
+				columnV.push_back(columnIndex); //Initially V=I -> 1's along diagonal
 
-				// //Try inserting other vertices into the simplex
-				// for(unsigned i=nPts; i-- != 0; ){ 
-					
-					
-				// 	if(it != simplex.first.rend() && i == *it - shift){ //Vertex i is already in the simplex
-				// 		//Now adding vertices less than i -> i is now the kth largest vertex in the simplex instead of the (k-1)th
-				// 		index -= bin.binom(i, k-1);
-				// 		index += bin.binom(i, k); //Recompute the index accordingly
-
-				// 		//Now need to check for the (k-1)th vertex in the simplex
-				// 		--k;
-						
-				// 		//Check for the previous vertex in the simplex (it is a reverse iterator)
-				// 		++it;
-				// 	} else{
-						
-						
-				// 		auto cofacetIndex = indexConverter.find(index + bin.binom(i, k));
-				// 		if(cofacetIndex != indexConverter.end()){ //If this is a valid simplex, add it to the heap
-				// 			_cofaceList.push_back(cofacetIndex->second);
-
-				// 			//If we haven't found an emergent candidate and the weight of the maximal cofacet is equal to the simplex's weight
-				// 			//		we have identified an emergent pair; at this point we can break because the interval is born and dies at the 
-				// 			//		same epsilon
-				// 			if(!foundEmergentCandidate && edges[d+1][cofacetIndex->second].second == simplex.second){
-								
-				// 				//Check to make sure the identified cofacet isn't a pivot
-				// 				if(_pivotPairs.find(cofacetIndex->second) == _pivotPairs.end()) 
-				// 					break; //Found an emergent cofacet pair -> we can break
-									
-				// 				foundEmergentCandidate = true;
-				// 			}
-				// 		}
-					// }
-				// }
-
-				//Never was able to iterate back to simplex.first.rend(); index holds the remapped coefficient
-
-
-				//		build a heap using the coface list to reduce and store in V
-				// std::make_heap(_cofaceList.begin(), _cofaceList.end(), std::greater<unsigned>());
+				//Build a heap using the coface list to reduce and store in V
 				std::make_heap(cofaceList.begin(), cofaceList.end(), cmpBySecond());
 
 				while(!cofaceList.empty()){
@@ -310,7 +187,7 @@ pipePacket fastPersistence::runPipe(pipePacket inData){
 						}
 
 						break;
-					} else{
+					} else{ //Reduce the column of R by computing the appropriate columns of D by enumerating cofacets
 						for(unsigned i : v[pivotPairs[pivot]]){
 							std::vector<simplexBase::treeNode*> cofaces = inData.complex->getAllCofacets(edges[d][i].first);
 							cofaceList.insert(cofaceList.end(), cofaces.begin(), cofaces.end());
@@ -324,85 +201,6 @@ pipePacket fastPersistence::runPipe(pipePacket inData){
 					bettiBoundaryTableEntry des = { d, simplexWeight, maxEpsilon, {} };
 					inData.bettiTable.push_back(des);
 				}
-
-				// while(!_cofaceList.empty()){
-					
-				// 	unsigned pivotIndex; //Get minimum coface from heap
-								
-					
-				// 	while(!_cofaceList.empty()){
-				// 		pivotIndex = _cofaceList.front();
-						
-				// 		//Rotate the heap
-				// 		std::pop_heap(_cofaceList.begin(), _cofaceList.end(), std::greater<unsigned>());
-						
-						
-				// 		_cofaceList.pop_back();
-				// 		if(!_cofaceList.empty() && pivotIndex == _cofaceList.front()){ //Coface is in twice -> evaluates to 0 mod 2
-							
-				// 			//Rotate the heap
-				// 			std::pop_heap(_cofaceList.begin(), _cofaceList.end(), std::greater<unsigned>());
-				// 			_cofaceList.pop_back();
-				// 		} else{
-				// 			_cofaceList.push_back(pivotIndex);
-							
-				// 			std::push_heap(_cofaceList.begin(), _cofaceList.end(), std::greater<unsigned>());
-				// 			break;
-				// 		}
-				// 	}
-
-				// 	if(_cofaceList.empty()){ //Column completely reduced
-				// 		break;
-				// 	} else if(_pivotPairs.find(pivotIndex) == _pivotPairs.end()){ //Column cannot be reduced
-				// 		nextPivots.push(pivotIndex);
-				// 		_pivotPairs.insert({pivotIndex, columnIndex});
-
-				// 		// std::sort(_columnV.begin(),_columnV.end());
-				// 		// auto it = _columnV.begin();
-				// 		// while(it != _columnV.end()){
-				// 		// 	if((it+1)!=_columnV.end() && *it==*(it+1)) ++it;
-				// 		// 	else{
-				// 		// 		v[columnIndex].push_back(*it);
-				// 		// 	}
-				// 		// 	++it;
-				// 		// }
-				// 		_v[columnIndex] = _columnV;
-
-				// 		if(edges[d][columnIndex].second != edges[d+1][pivotIndex].second){
-				// 			std::cout<<simplexWeight<<' '<<edges[d+1][pivotIndex].second<<'\n';
-				// 			bettiBoundaryTableEntry des = { d, simplexWeight, edges[d+1][pivotIndex].second, std::set<unsigned>(_cofaceList.begin(), _cofaceList.end()) };
-				// 			inData.bettiTable.push_back(des);
-				// 		}
-
-				// 		break;
-				// 	} else{
-				// 		for(long long index : _v[_pivotPairs[pivotIndex]]){ //Reduce the column of R by computing the appropriate columns of D by enumerating cofacets
-				// 			_columnV.push_back(index);
-				// 			std::vector<unsigned> facet = getVertices(index, d, nPts, bin); //Enumerate cofacets of this simplex like above
-				// 			std::vector<unsigned>::iterator it = facet.begin();
-				// 			unsigned k = facet.size() + 1;
-				// 			for(unsigned i = nPts; i-- != 0; ){ //Try inserting other vertices into the simplex
-				// 				if(it != facet.end() && i == *it){
-				// 					index -= bin.binom(i, k-1);
-				// 					index += bin.binom(i, k);
-				// 					--k;
-				// 					++it;
-				// 				} else{
-				// 					auto cofacetIndex = indexConverter.find(index + bin.binom(i, k));
-				// 					if(cofacetIndex != indexConverter.end()){ //If this is a valid simplex, add it to the heap
-				// 						_cofaceList.push_back(cofacetIndex->second);
-				// 					}
-				// 				}
-				// 			}
-				// 		}
-				// 		std::make_heap(_cofaceList.begin(), _cofaceList.end(), std::greater<unsigned>());
-				// 	}
-				// }
-				
-				// if(_cofaceList.empty()){
-				// 	bettiBoundaryTableEntry des = { d, simplexWeight, maxEpsilon, {} };
-				// 	inData.bettiTable.push_back(des);
-				// }
 			
 			//Was a pivot, skip the evaluation and queue next pivot
 			} else pivots.pop();
